moving md log and traj cleaning fn to mlp utils
